requirements.txt
appdirs==1.4.3
Django==1.9.12
django-appconf==1.0.2
django-imagekit==4.0.1
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.1.1
pyparsing==2.2.0
six==1.10.0
